try SRP single Responsibility Principle
